CTS: fix libcore connection tests
As the note mentions, they are intentionally trying to connect to server
that doesn't exist.  But on many wifi networks, there is some server
listening on port 80 for management, login, etc.  So changing the port
to one that is even less likely to be used makes this test more reliable.

Change-Id: Ice93f98ada51b0d75a05cfea9db76dc4636e0827