Change os.arch property for compatibility
Some apps don't like seeing e.g. "x86_64" in the os.arch Java property
where they were written to expect "i686". If ENV variable
OS_ARCH_OVERRIDE is set, set os.arch to the value of OS_ARCH_OVERRIDE
instead of the value in the uname machine field.

It would be cleaner to use an Android OS property, but those are not
available to the core Java libraries (android.os.SystemProperties
depends on this code, not the reverse).

Issue: AXIA-2911
Change-Id: I0079a1afb55dc2d81f0a32ab0616652d221515e5
Signed-off-by: Xiaobing Feng <xiaobing.feng@windriver.com>